+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLPeerChannel.java
@@ -8,6 +8,7 @@ import java.io.OutputStream;
 
 import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
 import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
 
@@ -19,7 +20,7 @@ public class TLPeerChannel extends TLAbsPeer {
 
     public static final int CONSTRUCTOR_ID = 0xbddde532;
 
-    protected int channelId;
+    protected long channelId;
 
     private final String _constructor = "peerChannel#bddde532";
 
@@ -32,7 +33,7 @@ public class TLPeerChannel extends TLAbsPeer {
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(channelId, stream);
+        writeLong(channelId, stream);
     }
 
     @Override
@@ -58,7 +59,7 @@ public class TLPeerChannel extends TLAbsPeer {
         return CONSTRUCTOR_ID;
     }
 
-    public int getChannelId() {
+    public long getChannelId() {
         return channelId;
     }
