+++ tl/src/main/java/com/github/badoualy/telegram/tl/api/TLPeerChat.java
@@ -8,6 +8,7 @@ import java.io.OutputStream;
 
 import static com.github.badoualy.telegram.tl.StreamUtils.readInt;
 import static com.github.badoualy.telegram.tl.StreamUtils.writeInt;
+import static com.github.badoualy.telegram.tl.StreamUtils.writeLong;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_CONSTRUCTOR_ID;
 import static com.github.badoualy.telegram.tl.TLObjectUtils.SIZE_INT32;
 
@@ -19,7 +20,7 @@ public class TLPeerChat extends TLAbsPeer {
 
     public static final int CONSTRUCTOR_ID = 0xbad0e5bb;
 
-    protected int chatId;
+    protected long chatId;
 
     private final String _constructor = "peerChat#bad0e5bb";
 
@@ -32,7 +33,7 @@ public class TLPeerChat extends TLAbsPeer {
 
     @Override
     public void serializeBody(OutputStream stream) throws IOException {
-        writeInt(chatId, stream);
+        writeLong(chatId, stream);
     }
 
     @Override
@@ -58,7 +59,7 @@ public class TLPeerChat extends TLAbsPeer {
         return CONSTRUCTOR_ID;
     }
 
-    public int getChatId() {
+    public long getChatId() {
         return chatId;
     }
